chore: utility to serve FlightDataTaskResult
- when the task result is ready and the do_get gets the task identifier via
  the ticket, there is only one correct way to pick the result for the task
  and send it out via Flight RPC
- the `do_get_task_result` encapsulates this data prep/serving logic
- refactored flexfuns to use it (the impl was picked from flexfun)

JIRA: CQ-670
